# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: SIDJUA REST API
  version: "1.0.1"
  description: |
    SIDJUA Free — Multi-Agent Governance Framework REST API.

    All endpoints (except `/api/v1/health`) require a Bearer token:
    ```
    Authorization: Bearer <token>
    ```

    Tokens carry scopes: `readonly`, `operator`, `admin`, `bootstrap`.
    Endpoints that mutate state require at least `operator` scope.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Local SIDJUA instance
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: "AUTH-001" }
            message: { type: string, example: "Unauthorized" }
            detail:  { type: string }
    AgentStatus:
      type: string
      enum: [stopped, starting, active, idle, stopping, error]
    AgentSummary:
      type: object
      required: [id, name, tier, division, status]
      properties:
        id:       { type: string }
        name:     { type: string }
        tier:     { type: integer, enum: [1, 2, 3] }
        division: { type: string }
        status:   { $ref: "#/components/schemas/AgentStatus" }
        model:    { type: string }
        provider: { type: string }
...
paths:
...
  /api/v1/agents:
    get:
      summary: List agents
      operationId: listAgents
      tags: [Agents]
      parameters:
        - in: query
          name: division
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: "#/components/schemas/AgentStatus" }
      responses:
        "200":
          description: Agent list
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents: { type: array, items: { $ref: "#/components/schemas/AgentSummary" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/agents/{id}:
    get:
      summary: Get agent
      operationId: getAgent
      tags: [Agents]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentSummary" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update agent
      operationId: updateAgent
      tags: [Agents]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:     { type: string }
                division: { type: string }
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgentSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{id}/start:
    post:
      summary: Start agent
      operationId: startAgent
      tags: [Agents]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Agent started
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/agents/{id}/stop:
    post:
      summary: Stop agent
      operationId: stopAgent
      tags: [Agents]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Agent stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─────────────────────────────────────────────────────────────────────────
  # Tasks
  # ─────────────────────────────────────────────────────────────────────────
...
components:
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }